+++ data/duke.txt
@@ -1,6 +1,7 @@
 D | 0 | return book | 2024-01-23T23:59
 E | 0 | prject | 2024-01-23T20:00 | 2024-01-23T21:00
 D | 0 | borrow book | 2024-01-22T23:59
-T | 1 | play games
+T | 0 | play games
 D | 1 | submit email | 2024-10-10T23:59
-E | 0 | job | 2024-10-01T00:00 | 2024-10-02T23:59
+E | 1 | job | 2024-10-01T00:00 | 2024-10-02T23:59
+D | 0 | homework | 2024-02-10T23:59
