+++ src/main/java/command/Command.java
@@ -1,3 +1,10 @@
+package command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 
@@ -9,7 +16,7 @@ public abstract class Command {
         isExit = false;
     }
 
-    public abstract void excuteCommand(TaskList tasks,Ui ui, Storage storage) throws DukeException;
+    public abstract void excuteCommand(TaskList tasks, Ui ui, Storage storage) throws DukeException;
 
     public void confirmExit() {
         this.isExit = true;
