+++ src/main/java/command/OtherCommand.java
@@ -1,3 +1,10 @@
+package command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 public class OtherCommand extends Command {
 
     @Override
