+++ src/main/java/duke/Action.java
@@ -1,3 +1,5 @@
+package duke;
+
 public enum Action {
     LIST,
     TODO,
