+++ src/main/java/task/ToDo.java
@@ -1,0 +1,15 @@
+package task;
+
+import task.Task;
+
+public class ToDo extends Task {
+
+    public ToDo(String description, boolean isDone) {
+        super(description, isDone);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]["+ super.getStatusIcon() +"] " + super.getDescription();
+    }
+}
