+++ text-ui-test/EXPECTED2.TXT
@@ -52,7 +52,7 @@
     ____________________________________________________________
 
     ____________________________________________________________
-      Invalid format for new Deadline!
+      Invalid format for new task.Deadline!
     ____________________________________________________________
 
     ____________________________________________________________
@@ -66,11 +66,11 @@
     ____________________________________________________________
 
     ____________________________________________________________
-      Invalid format for new Event!
+      Invalid format for new task.Event!
     ____________________________________________________________
 
     ____________________________________________________________
-      Invalid format for new Event!
+      Invalid format for new task.Event!
     ____________________________________________________________
 
     ____________________________________________________________
